Create README.md
The Story of Independence Day and America’s Birthday
Independence Day is the national holiday of the United States of America commemorating the signing of the Declaration of Independence by the Continental Congress on July 4, 1776, in Philadelphia, Pennsylvania. US250 is creating a musical event to celebrate our 250 nations Birthday will bring together far
reaching concepts of the world to date with
the best and brightest hopes for the future
in a celebration of the century, highlighted
with some of the best in Rock music today.
The  proposal will outline the US250 celebrations and events 
planned to mark this turning point in human 
history, and give an early look at the efforts
being mounted to make this national 
Celebration a truly  special event July 4, 2026. Get in on one of the hottest NFTs celebrating this historic event.